refactor: extract event handler logic and improve file naming
- Create handlers/review-action-handler.ts for business logic and event handling
- Rename admin-review-actions.ts to admin-review-listeners.ts for better clarity
- Simplify main file to focus purely on DOM event listener setup
- Reduce main file from 82 lines to 26 lines (68% reduction)
- Improve separation of concerns between event binding and business logic
- Update import path in index.ts to reflect new filename
